young_readers/requirements.txt
Django==1.7.4
Unipath==1.0
argparse==1.2.1
bottlenose==0.6.2
cssmin==0.2.0
dj-database-url==0.3.0
dj-static==0.0.6
django-pipeline==1.4.4.1
django-toolbelt==0.0.1
django-user-agents==0.3.0
djangorestframework==3.0.4
futures==2.2.0
gunicorn==19.2.0
ply==3.4
psycopg2==2.5.4
python-amazon-product-api==0.2.8
requests==2.5.1
slimit==0.8.1
static3==0.5.1
ua-parser==0.3.6
user-agents==0.3.2
wsgiref==0.1.2
xmltodict==0.9.2
